# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * WhatsApp Bot Hosting Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SystemInfo {
  nodeVersion: string;
  platform: string;
}

export type BotStatus = typeof BotStatus[keyof typeof BotStatus];


export const BotStatus = {
  stopped: 'stopped',
  starting: 'starting',
  running: 'running',
  error: 'error',
} as const;

export interface Bot {
  id: number;
  name: string;
  gitUrl: string;
  command: string;
  status: BotStatus;
  /** @nullable */
  pid?: number | null;
  autoRestart: boolean;
  /**
     * GitHub token for private repositories
     * @nullable
     */
  gitToken?: string | null;
  /**
     * Custom dependency install command (e.g. npm install --legacy-peer-deps)
     * @nullable
     */
  installCommand?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface BotInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  gitUrl: string;
  /** @minLength 1 */
  command: string;
  autoRestart?: boolean;
  /** GitHub token for private repositories */
  gitToken?: string;
  /** Dependency install command (default: npm install --legacy-peer-deps) */
  installCommand?: string;
}

export interface BotUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  command?: string;
  /** @minLength 1 */
  gitUrl?: string;
  autoRestart?: boolean;
  gitToken?: string;
  installCommand?: string;
}

export interface PullResult {
  output: string;
}

export interface TerminalInput {
  text: string;
}

export interface LogsResponse {
  logs: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}